Remove dead code and mark public types internal
