+++ src/GitLink/Extensions/BinaryReaderExtensions.cs
@@ -12,16 +12,9 @@ namespace GitLink
     using System.Text;
     using Catel;
 
-    public static class BinaryReaderExtensions
+    internal static class BinaryReaderExtensions
     {
-        public static Guid ReadGuid(this BinaryReader binaryReader)
-        {
-            Argument.IsNotNull(() => binaryReader);
-
-            return new Guid(binaryReader.ReadBytes(16));
-        }
-
-        public static string ReadCString(this BinaryReader binaryReader)
+        internal static string ReadCString(this BinaryReader binaryReader)
         {
             Argument.IsNotNull(() => binaryReader);
