+++ src/GitLink/Extensions/PdbExtensions.cs
@@ -14,9 +14,9 @@ namespace GitLink
     using Catel;
     using Pdb;
 
-    public static class PdbExtensions
+    internal static class PdbExtensions
     {
-        public static IEnumerable<string> FindMissingOrChangedSourceFiles(this PdbFile pdbFile)
+        internal static IEnumerable<string> FindMissingOrChangedSourceFiles(this PdbFile pdbFile)
         {
             Argument.IsNotNull(() => pdbFile);
 
@@ -25,7 +25,7 @@ namespace GitLink
                 string file = checksumInfo.Key;
                 byte[] expectedChecksum = checksumInfo.Value;
                 HashAlgorithm hasher = expectedChecksum.Length == 16 ? (HashAlgorithm)MD5.Create() : SHA1.Create();
-                byte[] actualChecksum = File.Exists(file) ? Crypto.HashFile(hasher, file) : null;
+                byte[] actualChecksum = File.Exists(file) ? HashFile(hasher, file) : null;
 
                 if (!AreEqualBuffers(expectedChecksum, actualChecksum))
                 {
@@ -34,7 +34,7 @@ namespace GitLink
             }
         }
 
-        public static IReadOnlyDictionary<string, byte[]> GetFilesAndChecksums(this PdbFile pdbFile)
+        internal static IReadOnlyDictionary<string, byte[]> GetFilesAndChecksums(this PdbFile pdbFile)
         {
             Argument.IsNotNull(() => pdbFile);
 
@@ -63,6 +63,14 @@ namespace GitLink
             }
 
             return results;
+        }
+
+        private static byte[] HashFile(HashAlgorithm ha, string file)
+        {
+            using (var fs = File.OpenRead(file))
+            {
+                return ha.ComputeHash(fs);
+            }
         }
 
         private static bool AreEqualBuffers(byte[] first, byte[] second)
