+++ src/GitLink/Helpers/PdbStrHelper.cs
@@ -10,11 +10,11 @@ namespace GitLink
     using Catel;
     using Catel.Logging;
 
-    public static class PdbStrHelper
+    internal static class PdbStrHelper
     {
         private static readonly ILog Log = LogManager.GetCurrentClassLogger();
 
-        public static void Execute(string pdbStrFileName, string projectPdbFile, string pdbStrFile)
+        internal static void Execute(string pdbStrFileName, string projectPdbFile, string pdbStrFile)
         {
             Argument.IsNotNullOrWhitespace(() => projectPdbFile);
             Argument.IsNotNullOrWhitespace(() => pdbStrFile);
